Move pickWallet to WalletListModal.tsc
